requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-crontab==0.7.1
django-environ==0.12.0
djangorestframework==3.15.2
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.1
